Tab/Doc/Mem
Worked on how tab is handled when added to the list, if ch arg is equal to '\t' we must always move x equal to _tabSize. Documentations was also added to describe the purpose of each function in editorMode.c. A general function was implimented to assert the allocations made in editorMode.c -> this should probably be used in general in this sofwar